+++ src/main/java/com/chillibits/particulatematterapi/repository/main/SensorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface SensorRepository extends MongoRepository<Sensor, Integer> {
+public interface SensorRepository extends MongoRepository<Sensor, Long> {
     //Integer updateSensor(int chipId, double latitude, double longitude, double lastValueP1, double lastValueP2);
     List<RankingItem> getRankingByCity(int items);
     List<RankingItem> getRankingByCountry(int items);
